Update README.md
Deregistration The Financial Private Sector . 
be25e189-586f-4d6b-b4d8-97f67d2ffd04
warith akbar federal government justice civilian retirement trust partnership. We have a invested partnership in the United States economy. The the people have a contract with the banks and Private sector. To invest into market and give the investors a return on that investment. For President Trump be the president and be the best president ever and make the congress honest again. And to deregulated Private sector energy markets. To give the federal bill of rights second bill of rights equal rights equal wage . A trade market that give equal value to all investors. To station and tariff and market that profits of crisis that happens in United States.  
For president of United States to finish what Fredrick R dream .