The deaclation of US foreign peace policy. Ovel office Civilian Bill of rights . LICENSE.txt
Draft AUTHORS warithuddin Akbar 
United States Supreme Cammander Civil Cammand Economic and infrastructure Reform Draft Bill stage one .
Stage to full support and endorsement commitment to protect are government. For the government is part of the people.  And we understand that we shell give support to this gov in so this government will give support in are financial market and energy structure. If a fire are floods to flood are community that are government will seek justice or army Corp of engineers to be sent not money. As Washington DC can see that politics forge fires . That fires and flood funds private sector investment . The only way to have control of this is do send federal fire Marshall with Flood Marshall engineers . In Washington state that's about 400 miles from CA does not have mass fires that burn down are towns. We will evolve to are next generation without restriction and politics.